# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Leasing Operations Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /properties:
    get:
      tags: [Properties]
      operationId: listProperties
      summary: List properties for the account
      responses:
        '200':
          description: List of properties
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertyList'
        '401':
          description: Unauthorized
    post:
      tags: [Properties]
      operationId: createProperty
      summary: Create a new property
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePropertyBody'
      responses:
        '201':
          description: Property created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (admin only)
  /properties/{id}:
    patch:
      tags: [Properties]
      operationId: updateProperty
      summary: Update a property
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePropertyBody'
      responses:
        '200':
          description: Property updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '404':
          description: Not found

  # Twilio Numbers
  /twilio-numbers:
    get:
      tags: [TwilioNumbers]
      operationId: listTwilioNumbers
      summary: List Twilio numbers for the account
      responses:
        '200':
          description: List of Twilio numbers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TwilioNumberList'
    post:
      tags: [TwilioNumbers]
      operationId: createTwilioNumber
      summary: Add a Twilio number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTwilioNumberBody'
      responses:
        '201':
          description: Number added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TwilioNumber'
  /twilio-numbers/{id}:
    patch:
      tags: [TwilioNumbers]
      operationId: updateTwilioNumber
      summary: Update a Twilio number
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTwilioNumberBody'
      responses:
        '200':
          description: Number updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TwilioNumber'

  # Users
...
  /prospects:
    get:
      tags: [Prospects]
      operationId: listProspects
      summary: List prospects
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: exportStatus
          in: query
          schema:
            type: string
        - name: propertyId
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Prospect list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProspectList'
    post:
      tags: [Prospects]
      operationId: createProspect
      summary: Create a new prospect
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProspectBody'
      responses:
        '201':
          description: Prospect created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prospect'
  /prospects/{id}:
    get:
      tags: [Prospects]
      operationId: getProspect
      summary: Get a single prospect with interactions and notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Prospect detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProspectDetail'
        '404':
          description: Not found
    patch:
      tags: [Prospects]
      operationId: updateProspect
      summary: Update a prospect
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProspectBody'
      responses:
        '200':
          description: Prospect updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prospect'
  /prospects/{id}/notes:
    post:
      tags: [Prospects]
      operationId: addProspectNote
      summary: Add a note to a prospect
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddNoteBody'
      responses:
        '201':
          description: Note added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
  /prospects/{id}/tags:
    post:
      tags: [Prospects]
      operationId: setProspectTags
      summary: Set tags on a prospect (replaces existing)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetTagsBody'
      responses:
        '200':
          description: Tags updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagList'

  # Tags
...
components:
...
  schemas:
...
    Property:
      type: object
      required: [id, accountId, name, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        accountId:
          type: string
        name:
          type: string
        address1:
          type: [string, 'null']
        address2:
          type: [string, 'null']
        city:
          type: [string, 'null']
        state:
          type: [string, 'null']
        zip:
          type: [string, 'null']
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PropertyList:
      type: object
      required: [properties]
      properties:
        properties:
          type: array
          items:
            $ref: '#/components/schemas/Property'
    CreatePropertyBody:
      type: object
      required: [name]
      properties:
        name:
          type: string
        address1:
          type: string
        address2:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        status:
          type: string
    UpdatePropertyBody:
      type: object
      properties:
        name:
          type: string
        address1:
          type: string
        address2:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        status:
          type: string

    # TwilioNumber schemas
    TwilioNumber:
      type: object
      required: [id, accountId, phoneNumber, isActive, createdAt, updatedAt]
      properties:
        id:
          type: string
        accountId:
          type: string
        propertyId:
          type: [string, 'null']
        phoneNumber:
          type: string
        friendlyName:
          type: [string, 'null']
        purpose:
          type: [string, 'null']
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TwilioNumberList:
      type: object
      required: [twilioNumbers]
      properties:
        twilioNumbers:
          type: array
          items:
            $ref: '#/components/schemas/TwilioNumber'
    CreateTwilioNumberBody:
      type: object
      required: [phoneNumber]
      properties:
        phoneNumber:
          type: string
        friendlyName:
          type: string
        propertyId:
          type: string
        purpose:
          type: string
    UpdateTwilioNumberBody:
      type: object
      properties:
        friendlyName:
          type: string
        propertyId:
          type: string
        purpose:
          type: string
        isActive:
          type: boolean

    # User schemas
...
    Prospect:
      type: object
      required: [id, accountId, phonePrimary, status, exportStatus, createdAt, updatedAt]
      properties:
        id:
          type: string
        accountId:
          type: string
        assignedPropertyId:
          type: [string, 'null']
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        fullName:
          type: [string, 'null']
        phonePrimary:
          type: string
        phoneSecondary:
          type: [string, 'null']
        email:
          type: [string, 'null']
        desiredMoveInDate:
          type: [string, 'null']
        desiredBedrooms:
          type: [string, 'null']
        budgetMin:
          type: [number, 'null']
        budgetMax:
          type: [number, 'null']
        pets:
          type: [string, 'null']
        voucherType:
          type: [string, 'null']
        employmentStatus:
          type: [string, 'null']
        monthlyIncome:
          type: [number, 'null']
        languagePreference:
          type: [string, 'null']
        latestSummary:
          type: [string, 'null']
        latestSentiment:
          type: [string, 'null']
        qualificationScore:
          type: [number, 'null']
        status:
          type: string
        exportStatus:
          type: string
        crmExternalId:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProspectList:
      type: object
      required: [prospects, total]
      properties:
        prospects:
          type: array
          items:
            $ref: '#/components/schemas/Prospect'
        total:
          type: integer
    ProspectDetail:
      type: object
      required: [prospect]
      properties:
        prospect:
          $ref: '#/components/schemas/Prospect'
        interactions:
          type: array
          items:
            $ref: '#/components/schemas/Interaction'
        notes:
          type: array
          items:
            $ref: '#/components/schemas/Note'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
    CreateProspectBody:
      type: object
      required: [phonePrimary]
      properties:
        phonePrimary:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        assignedPropertyId:
          type: string
        status:
          type: string
    UpdateProspectBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        fullName:
          type: string
        email:
          type: string
        phoneSecondary:
          type: string
        assignedPropertyId:
          type: string
        desiredMoveInDate:
          type: string
        desiredBedrooms:
          type: string
        budgetMin:
          type: number
        budgetMax:
          type: number
        pets:
          type: string
        voucherType:
          type: string
        employmentStatus:
          type: string
        monthlyIncome:
          type: number
        languagePreference:
          type: string
        status:
          type: string
        exportStatus:
          type: string
        crmExternalId:
          type: string

    # Note schemas
    Note:
      type: object
      required: [id, accountId, prospectId, userId, body, createdAt]
      properties:
        id:
          type: string
        accountId:
          type: string
        prospectId:
          type: string
        userId:
          type: string
        body:
          type: string
        createdAt:
          type: string
          format: date-time
    AddNoteBody:
      type: object
      required: [body]
      properties:
        body:
          type: string

    # Tag schemas
    Tag:
      type: object
      required: [id, accountId, name]
      properties:
        id:
          type: string
        accountId:
          type: string
        name:
          type: string
        color:
          type: [string, 'null']
    TagList:
      type: object
      required: [tags]
      properties:
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
...
    SetTagsBody:
      type: object
      required: [tagIds]
      properties:
        tagIds:
          type: array
          items:
            type: string

    # Interaction schemas
    Interaction:
      type: object
      required: [id, accountId, sourceType, direction, fromNumber, toNumber, occurredAt, createdAt, updatedAt]
      properties:
        id:
          type: string
        accountId:
          type: string
        prospectId:
          type: [string, 'null']
        propertyId:
          type: [string, 'null']
        sourceType:
          type: string
        direction:
          type: string
        twilioMessageSid:
          type: [string, 'null']
        twilioCallSid:
          type: [string, 'null']
        parentThreadKey:
          type: [string, 'null']
        fromNumber:
          type: string
        toNumber:
          type: string
        rawText:
          type: [string, 'null']
        transcript:
          type: [string, 'null']
        summary:
          type: [string, 'null']
        category:
          type: [string, 'null']
        urgency:
          type: [string, 'null']
        sentiment:
          type: [string, 'null']
        extractionConfidence:
          type: [number, 'null']
        structuredExtractionJson:
          type: [object, 'null']
          additionalProperties: true
        extractionStatus:
          type: [string, 'null']
        occurredAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...