# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Leasing Operations Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface Property {
  id: string;
  accountId: string;
  name: string;
  /** @nullable */
  address1?: string | null;
  /** @nullable */
  address2?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  zip?: string | null;
  status: string;
  createdAt: Date;
  updatedAt: Date;
}