# lib/api-zod/src/generated/types/twilioNumberList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Leasing Operations Panel API
 * OpenAPI spec version: 0.1.0
 */
import type { TwilioNumber } from "./twilioNumber";

export interface TwilioNumberList {
  twilioNumbers: TwilioNumber[];
}